Issue #39: No longer having separate copy of Google Play library
Got rid of the separate copy for command line operation.  It's not
necessary; the issue it was trying to address was a problem where
cleaning from the command line has eclipse confused.  First, cleaning
probably won't be a common event, and second, having eclipse clean as
well solves the problem.