Moving state out of fragments
Added IRBuddyActivity, IRBuddyActivityListener; also realizing that
Fragments should be considered views within the model/view/controller
paradigm, and NOT states, since their lifespan is unpredictable so they
can't reliably store state of interest to other application components.
For instance, the 'active receipt' should be stored with the
RBuddyActivity, NOT in the ReceitEditor fragment.